refactor : MainScreen 분리
